+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/b_commands/auto/ParkCommandWithout.java
@@ -61,7 +61,7 @@ public class ParkCommandWithout extends CommandBase {
 				            .build();
 			default:
 				return drive.trajectorySequenceBuilder(startPose)
-				            .forward(1)
+				            .back(0.01)
 				            .build();
 		}
 	}
